multi-arfcn, trx: scale receive samples to expected range
Scale the +/- 1.0 floating point input range to a range
of +/- 9450.0. The UHD device was converted to output
floating point samples in the commit:

"multi-arfcn, uhd: output floating point directly from uhd"

The result was the input sample range was reduced from the
device approximately +/- 32000 integer to +/- 1.0 floating
point. Integer samples were converted to floating point
directly, which created a discrepency between expected
sample amplitude being fed into the transceiver.

The exepected "full scale" range is actually less than the
full signed 16-bit range. A value of 9450.0 is given in
USRPDevice.h; the specific reason for this value is unknown.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>