multi-arfcn, uhd: make tx amplitude dependent on number of ARFCN's
The default amplitude setting will cause the signal to splatter
when the number of ARFCN's used is high. This occurs because the
amplitude is set for a single channel only and does not account
for the composite signal at the output of the synthesis filter.

The previous default value - 30% of full dynamic range - was
likely too low given that constant GMSK is used. Increase the
single channel value to 50%. In the multi-ARFCN case, divide
this amplitude setting by the number ARFCN being used.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>