transceiver: mcbts:  use matching filter lengths for internal resampler
Remove the option for separate filter lengths between the channelizing
and resampling filterbanks. We need this change in order to setup
in-place FFTW transforms - which allows removal of interleavering
in to and out of the FFT calls. In the process, we can use shorter
variable names since we don't need to distinguish between resampling
and channelizing filter lengths.

Signed-off-by: Thomas Tsou <tom@tsou.cc>